update 3 function 0-print_listint

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,14 +1,15 @@
 #include "lists.h"
 
 /**
- * print_listint - print all the elements of listint_t
- * @h: head pointer
- * Return: number of nodes
+ * print_listint - prints all the elements of a
+ * listint_t list.
+ * @h: head
+ * Return: print element and return number
  */
 
 size_t print_listint(const listint_t *h)
 {
-	size_t nodeNumber = 0;
+	unsigned int nodeNumber = 0;
 
 	while (h != NULL)
 	{
